+++ packages/dialog/src/dialog.js
@@ -1,6 +1,5 @@
 import Vue from 'vue';
 import Dialog from './dialog.vue';
-import merge from 'src/utils/merge';
 
 const DialogConstructor = Vue.extend(Dialog);
 
@@ -24,6 +23,9 @@ const initInstance = () => {
     el: document.createElement('div')
   });
 
+  instance.$on('input', value => {
+    instance.value = value;
+  })
   instance.callback = defaultCallback;
 };
 
@@ -36,7 +38,7 @@ const showNextDialog = () => {
   if (!instance.value && dialogQueue.length > 0) {
     currentDialog = dialogQueue.shift();
 
-    const options = currentDialog.options;
+    const { options } = currentDialog;
 
     for (const prop in options) {
       /* istanbul ignore else */
@@ -45,22 +47,16 @@ const showNextDialog = () => {
       }
     }
 
-    if (options.callback === undefined) {
-      instance.callback = defaultCallback;
-    }
-
+    instance.callback = options.callback || defaultCallback;
+    instance.value = true;
     document.body.appendChild(instance.$el);
-
-    Vue.nextTick(() => {
-      instance.value = true;
-    });
   }
 };
 
 var DialogBox = options => {
   return new Promise((resolve, reject) => { // eslint-disable-line
     dialogQueue.push({
-      options: merge({}, options),
+      options: { ...options },
       callback: options.callback,
       resolve: resolve,
       reject: reject
@@ -71,23 +67,25 @@ var DialogBox = options => {
 };
 
 DialogBox.alert = function(options) {
-  return DialogBox(merge({
+  return DialogBox({
     type: 'alert',
     title: '',
     message: '',
     closeOnClickOverlay: false,
-    showCancelButton: false
-  }, options));
+    showCancelButton: false,
+    ...options
+  });
 };
 
 DialogBox.confirm = function(options) {
-  return DialogBox(merge({
+  return DialogBox({
     type: 'confirm',
     title: '',
     message: '',
     closeOnClickOverlay: true,
-    showCancelButton: true
-  }, options));
+    showCancelButton: true,
+    ...options
+  });
 };
 
 DialogBox.close = function() {