+++ packages/steps/src/steps.vue
@@ -24,7 +24,7 @@
 </template>
 
 <script>
-import Icon from 'packages/icon';
+import Icon from '../../icon';
 
 export default {
   name: 'van-steps',